Change symbols position field to Long, matching change in JFlex

+++ src/main/java/com/bertramlabs/plugins/hcl4j/RuntimeSymbols/BooleanPrimitiveType.java
@@ -1,7 +1,7 @@
 package com.bertramlabs.plugins.hcl4j.RuntimeSymbols;
 
 public class BooleanPrimitiveType extends PrimitiveType {
-	public BooleanPrimitiveType(Integer line, Integer column,Integer position) {
+	public BooleanPrimitiveType(Integer line, Integer column, Long position) {
 		super("boolean",line,column,position);
 	}
