+++ src/main/java/com/bertramlabs/plugins/hcl4j/RuntimeSymbols/NumberPrimitiveType.java
@@ -1,7 +1,7 @@
 package com.bertramlabs.plugins.hcl4j.RuntimeSymbols;
 
 public class NumberPrimitiveType extends PrimitiveType {
-	public NumberPrimitiveType(Integer line, Integer column,Integer position) {
+	public NumberPrimitiveType(Integer line, Integer column, Long position) {
 		super("number",line,column,position);
 	}
